% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{analyze_popularity}
\alias{analyze_popularity}
\title{2. Analyze Predictors for POPULARITY (Regression)}
\usage{
analyze_popularity(data)
}
\arguments{
\item{data}{The cleaned dataset}
}
\description{
2. Analyze Predictors for POPULARITY (Regression)
}
