% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{prepare_ml_data}
\alias{prepare_ml_data}
\title{1. Data Preparation & Sampling
2.1 Million rows is very large for local ML. We will sample it.}
\usage{
prepare_ml_data(df)
}
\arguments{
\item{df}{The spotify_processed dataframe}
}
\description{
1. Data Preparation & Sampling
2.1 Million rows is very large for local ML. We will sample it.
}
